!!!I'm pleased to inform that the program is WORKING!!!
Add, Remove, Update are working

Still need to do some tweaks related with actions performed outside the database, for instance: open, save and export files!

For now, to import a file please be sure that is a XML file with the structure specified in the exemples/base.xml. Speaking of which, the program starts with the sentences imported from this file.